[modify]分享日志的逻辑转移至CrashHelper，提供给外部调用
